Make the "tab-as-character" behaviour opt-in.
Added `tab_moves_focus()` to `TextEdit` so we can opt-in to the
"tab-as-character" behaviour.

Also because we decided to lock the tab size to 4 spaces, we gave
up on the `tab_size` property, but we kept the opt-in for the
"tab-as-spaces" behaviour.

An updated example of how to use `TextEdit` with those two opt-in
behaviours can be found in the "Misc Demos" demo.