added support for fullscreen window mode
The default state can be set with the `NativeOptions` at compile time,
and can be changed by using a callback to the UI context during runtime.
The fullscreen state is written to the RON file so that the user-set
state is restored on each launch.
I added a 'toggle fullscreen' button to the UI too, so that can be
previewed.