Increase scroll speed for glium backend
Each so called "line" of the MuseScrollDelta::LineDelta de-facto represents
a single scroller click which in most cases equals to 15 degrees of wheel
rotation.

Currently, one scroller click scrolls exactly one line of the text,
which is very slow.

This change multiplies the scroll rate by 5. So, with one click of the
scroller, 5 lines scrolled. This number is taken from the "iced" GUI
library and feels quite balanced.

Fixes #461