egui/memory: expose Area State to users
The reason I want to to make these things publicly accessible is that I want to be able to retrieve the state of an area by its Id so that I can access that area's bounding rectangle after it's been rendered in order to pull just that part of the gui app out of the GPU for the sake of sending that part of the app to Open Broadcast Software while streaming.

The practical use case of this is exposing only an inner window containing a chatbox in my stream without also exposing stream controls. I can image similar use cases for other interesting gui elements that I'd like to implement in the future. If it's helpful, I can push my motivating use case to a branch on my twitch bot repo.

I don't know if this is the best way to do what I want, so I'm definitely open to alternative approaches. I made these changes way back in April last time I was working on my dumpy twitch bot. I may have been pointed in this direction by someone in Discord, cause I have no recollection of how I figured this out.