`eframe`: selectively expose parts of the API based on compile target
A lot of the `eframe` API is native-only or web-only. With this PR,
only the parts that are implemented for each platform is exposed.

This means you'll need to add `#[cfg(target_arch = "wasm32")]` around
code that uses the web-parts of the eframe API, and add
`#[cfg(not(target_arch = "wasm32"))]` around the parts that are
for native/desktop.