feat: provide option to hide scroll bar unless interacting with the scroll area
The opacity of the scroll bar may optionally be tied
to the interaction the user is doing with the scroll area.

This has prior art in things like VSCode, which don't show scroll bars
on containers unless you're hovered over them or interacting
with the scroll bar.

This effect is disabled by default.