eframe web: Don't throw away frames on click/copy/cut
To work around a Safari limitation, we run the app logic in the event
handler of copy, cut, and mouse up and down.

Previously the output of that frame was discarded, but in this PR
it is now saved to be used in the next requestAnimationFrame.

The result is noticeable more distinct clicks on buttons
(one more frame of highlight)