egui-winit: Handle `Cut`, `Copy` and `Paste` commands
The comment added in commit 8a0bc97e ("[egui_glium] Fix paste") seems
to assume that `winit` "should have translated" common "paste" keyboard
combos to a `Cut`/`Copy`/`Paste` "KeyCode", but completely glossed over
the fact that this `KeyCode` (now also `NamedKey`) maps to a special
key dedicated to this purpose found on some keyboards and OSes.  Make
sure that this key is still handled in addition to the combo that is
detected.